refactor(pointer-scanner): unify state machine and add detach safety
- Replace multiple atomic bools with single ScannerState enum
- Add HasPendingResult/ApplyPendingResult/Update for explicit async handling
- Add ProcessWillDetach signal to safely cancel operations before detach
- Optimize module lookup with binary search in pointer_scanner.cpp
- Modernize core with std::ranges and designated initializers
- Enhance lint.py with MSVC toolset detection and Windows VT support

+++ src/maia/core/process.cpp
@@ -3,6 +3,7 @@
 #include "maia/core/process.h"
 
 #include <algorithm>
+#include <cstddef>
 #include <optional>
 
 #include "maia/core/memory_common.h"
@@ -172,7 +173,7 @@ bool Process::ReadMemory(std::span<const MemoryAddress> addresses,
     return true;
   }
 
-  constexpr size_t kMaxBatchSize = 64 * 1024;
+  constexpr size_t kMaxBatchSize = static_cast<const size_t>(64 * 1024);
   constexpr size_t kMaxGapBytes = 256;
 
   auto indexed_addresses = CreateIndexedAddresses(addresses);
